+++ memcheck/vignettes.R
@@ -1,0 +1,1 @@
+pkgdown::build_articles(preview = FALSE)